Acrobat_Control_with_Coarse_Coding_RL/code/requirements.txt
numpy
matplotlib
prettytable
tensorflow
pyyaml